+++ apps/storybook/src/orderlyConfig/dataAdapter.ts
@@ -20,22 +20,6 @@ export const dataAdapter: OrderlyAppProviderProps["dataAdapter"] = {
         type: AnnouncementType.Campaign,
         coverImage: "https://i.ibb.co/HY0QfQT/Image-12.webp",
       },
-      {
-        announcement_id: "222",
-        message:
-          "DAWN OF DOMINANCE: $25,000 Trading Campaign is live Trading Campaign is live 222",
-        url: "https://app.orderly.network/tradingRewards",
-        type: AnnouncementType.Campaign,
-        coverImage: "https://i.ibb.co/HY0QfQT/Image-12.webp",
-      },
-      {
-        announcement_id: "333",
-        message:
-          "DAWN OF DOMINANCE: $25,000 Trading Campaign is live Trading Campaign is live 333",
-        url: "https://app.orderly.network/tradingRewards",
-        type: AnnouncementType.Campaign,
-        coverImage: "https://i.ibb.co/HY0QfQT/Image-12.webp",
-      },
       ...data,
     ];
   },