+++ packages/ui-scaffold/src/components/main/mainMenus/navItem.tsx
@@ -48,6 +48,7 @@ export type MainNavItem = {
   id?: string;
   testid?: string;
   name: string;
+  label?: string;
   href: string;
   target?: HTMLAttributeAnchorTarget;
   icon?: string | React.ReactElement;
@@ -180,6 +181,7 @@ export const NavItem: FC<
         data-actived={isActive}
         className={cn(
           "oui-group oui-relative oui-h-[32px] oui-rounded oui-px-3 oui-py-1 oui-text-sm oui-text-base-contrast-36 hover:oui-bg-base-7",
+          item.className,
           classNames?.navItem,
         )}
         onClick={onClickHandler}
@@ -191,7 +193,7 @@ export const NavItem: FC<
             angle={45}
             className="oui-whitespace-nowrap oui-break-normal"
           >
-            {item.name}
+            {item.label ?? item.name}
           </Text.gradient>
           {hasSubMenu && (
             <span className={"oui-ml-1 group-data-[open=true]:oui-rotate-180"}>
@@ -327,7 +329,7 @@ const SubMenus: React.FC<
         className={cn(
           customSubMenuRender
             ? "oui-w-auto oui-p-0 oui-border-0 oui-rounded-lg"
-            : "oui-w-[260px] oui-space-y-[2px] oui-border oui-border-line-6 oui-p-1",
+            : "oui-w-[200px] oui-space-y-[2px] oui-border oui-border-line-6 oui-p-1",
           className,
         )}
       >
@@ -414,7 +416,7 @@ const SubMenuTitle: React.FC<{ item: MainNavItem; isActive?: boolean }> = (
         </Text.gradient>
         {typeof item.tag !== "undefined" && <Tag item={item} />}
       </div>
-      {item.target === "_blank" && <OutlinkIcon />}
+      {/* {item.target === "_blank" && <OutlinkIcon />} */}
     </Flex>
   );
 };