+++ packages/web3-provider-ethers/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @orderly.network/web3-provider-ethers
 
+## 2.8.3
+
+### Patch Changes
+
+- publish
+- 2d3313c: publish
+- Updated dependencies
+- Updated dependencies [2d3313c]
+  - @orderly.network/default-evm-adapter@2.8.3
+  - @orderly.network/types@2.8.3
+
+## 2.8.3-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/default-evm-adapter@2.8.3-alpha.0
+  - @orderly.network/types@2.8.3-alpha.0
+
 ## 2.8.2
 
 ### Patch Changes